refactor: isolate refresh token in custom hook
