+++ src/app/user/[id]/page.tsx
@@ -1,25 +1,13 @@
 'use client'
 
-import refreshToken from '@/actions/refresh-token'
-import { useSearchParams } from 'next/navigation'
-import { useEffect } from 'react'
+import useRefreshToken from '@/hooks/useRefreshToken'
 
 /**
  * User page component.
  * @returns The page component.
  */
 export default function Page() {
-	const searchParams = useSearchParams()
-
-	useEffect(() => {
-		async function refreshTokenWrapper() {
-			await refreshToken()
-		}
-
-		if (!searchParams.get('recentToken')) {
-			refreshTokenWrapper()
-		}
-	}, [searchParams])
+	useRefreshToken()
 
 	return (
 		<main>
